[2573] return the original double of progress, instead of convert to int.
so the caller can get highest possible precission.

+++ src/lib/datasrc/zone_loader.h
@@ -150,23 +150,23 @@ public:
     /// \throw None
     size_t getRRCount() const;
 
-    /// \brief Return the current progress of the loader in percentage.
+    /// \brief Return the current progress of the loader.
     ///
-    /// This method returns the current estimated progress of loader in
-    /// percentage; it's 0 before starting the load, and 100 at the
-    /// completion, and a value between 0 and 100 in the middle of loading.
-    /// It's an implementation detail how to calculate the progress, which
-    /// may vary depending on how the loader is constructed and may even be
-    /// impossible to detect effectively.
+    /// This method returns the current estimated progress of loader as a
+    /// value between 0 and 1 (inclusive); it's 0 before starting the load,
+    /// and 1 at the completion, and a value between these (exclusive) in the
+    /// middle of loading.  It's an implementation detail how to calculate
+    /// the progress, which may vary depending on how the loader is
+    /// constructed and may even be impossible to detect effectively.
     ///
     /// If the progress cannot be determined, this method returns a special
     /// value of PROGRESS_UNKNOWN, which is not included in the range between
-    /// 0 and 100.
+    /// 0 and 1.
     ///
     /// As such, the application should use the return value only for
     /// informational purposes such as logging.  For example, it shouldn't
     /// be used to determine whether loading is completed by comparing it
-    /// to 100.  It should also expect the possibility of getting
+    /// to 1.  It should also expect the possibility of getting
     /// \c PROGRESS_UNKNOWN at any call to this method; it shouldn't assume
     /// the specific way of internal implementation as described below (which
     /// is provided for informational purposes only).
@@ -187,13 +187,13 @@ public:
     /// source can provide the latter value efficiently.
     ///
     /// \throw None
-    int getProgress() const;
+    double getProgress() const;
 
     /// \brief A special value for \c getProgress, meaning the progress is
     /// unknown.
     ///
     /// See the method description for details.
-    static const int PROGRESS_UNKNOWN;
+    static const double PROGRESS_UNKNOWN;
 
 private:
     /// \brief The iterator used as source of data in case of the copy mode.
